Fix swift

+++ LLDebugTool/Helper/LogHelper/LLLogHelper.h
@@ -57,6 +57,6 @@
  @param onEvent Event,can filter by this.
  @param message Message.
  */
-- (void)logInFile:(NSString *)file function:(NSString *)function lineNo:(int)lineNo level:(LLConfigLogLevel)level onEvent:(NSString *)onEvent message:(NSString *)message;
+- (void)logInFile:(NSString *)file function:(NSString *)function lineNo:(NSInteger)lineNo level:(LLConfigLogLevel)level onEvent:(NSString *)onEvent message:(NSString *)message;
 
 @end
